Move PurchaseConfig to its own file
